// src/modules/ai/ai-urls.enum.ts
export enum AI_URLS_KEYS {
  Dermatology = 'DERMALTOLOGY_URL',
  Retinopathy = 'RETINOPATHY_URL',
  DurationofDiabetes = 'DIABETES_DURATION_URL',
  ID_Scan = 'ID_SCAN_URL',
  TranslateToEn = 'TRANSLATE_TO_ENG_URL',
  TranslateToAr = 'TRANSLATE_TO_AR_URL',
  Scan = 'SCAN_URL',
  CovidScan = 'COVID_CLASSIFY_URL',
  Chest = 'CHEST_URL',
  StrokeHemorrhage = 'Stroke_Hemorrhage_URL',
  StrokeLocal = 'Stroke_Local_URL',
  SegmentationPredict = 'SEGMENTATION_PREDICT_URL',
  UpperBonesPredict = 'UpperBones_PredictUrl',
  LowerBonesPredict = 'LowerBones_PredictUrl',
  ICD10_Text = 'ICD10_Text',
  ICD10_Code = 'ICD10_Code',
  ICD10_ExtractSymptoms = 'ICD10_ExtractSymptoms',
  ICD10_SymptomsCoding = 'ICD10_SymptomsCoding',
}